Chore: Refactor request processors into separate package and individual files.
This sets up a place to create new request and response processors for e.g. context management.
