+++ core/src/agents/llm_agent.ts
@@ -4,58 +4,32 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import {
-  Content,
-  FunctionCall,
-  GenerateContentConfig,
-  Part,
-  Schema,
-} from '@google/genai';
+import {GenerateContentConfig, Schema} from '@google/genai';
 import {context, trace} from '@opentelemetry/api';
-import {cloneDeep} from 'lodash-es';
-import {z} from 'zod';
+
 import {z as z3} from 'zod/v3';
 import {z as z4} from 'zod/v4';
 
-import {
-  BaseCodeExecutor,
-  isBaseCodeExecutor,
-} from '../code_executors/base_code_executor.js';
-import {isBuiltInCodeExecutor} from '../code_executors/built_in_code_executor.js';
-import {
-  buildCodeExecutionResultPart,
-  buildExecutableCodePart,
-  CodeExecutionResult,
-  convertCodeExecutionParts,
-  extractCodeAndTruncateContent,
-  File,
-} from '../code_executors/code_execution_utils.js';
-import {CodeExecutorContext} from '../code_executors/code_executor_context.js';
+import {BaseCodeExecutor} from '../code_executors/base_code_executor.js';
+
 import {
   createEvent,
   createNewEventId,
   Event,
   getFunctionCalls,
-  getFunctionResponses,
   isFinalResponse,
 } from '../events/event.js';
-import {createEventActions} from '../events/event_actions.js';
+
 import {BaseExampleProvider} from '../examples/base_example_provider.js';
 import {Example} from '../examples/example.js';
 import {BaseLlm, isBaseLlm} from '../models/base_llm.js';
-import {
-  appendInstructions,
-  LlmRequest,
-  setOutputSchema,
-} from '../models/llm_request.js';
+import {LlmRequest} from '../models/llm_request.js';
 import {LlmResponse} from '../models/llm_response.js';
 import {LLMRegistry} from '../models/registry.js';
-import {State} from '../sessions/state.js';
+
 import {BaseTool, isBaseTool} from '../tools/base_tool.js';
 import {BaseToolset} from '../tools/base_toolset.js';
-import {FunctionTool} from '../tools/function_tool.js';
-import {ToolConfirmation} from '../tools/tool_confirmation.js';
-import {base64Decode} from '../utils/env_aware_utils.js';
+
 import {logger} from '../utils/logger.js';
 import {Context} from './context.js';
 
@@ -69,22 +43,24 @@ import {BaseAgent, BaseAgentConfig} from './base_agent.js';
 import {
   BaseLlmRequestProcessor,
   BaseLlmResponseProcessor,
-} from './base_llm_processor.js';
-import {
-  getContents,
-  getCurrentTurnContents,
-} from './content_processor_utils.js';
+} from './processors/base_llm_processor.js';
+
 import {
   generateAuthEvent,
   generateRequestConfirmationEvent,
   getLongRunningFunctionCalls,
-  handleFunctionCallList,
   handleFunctionCallsAsync,
   populateClientFunctionCallId,
-  REQUEST_CONFIRMATION_FUNCTION_CALL_NAME,
 } from './functions.js';
-import {injectSessionState} from './instructions.js';
+
 import {InvocationContext} from './invocation_context.js';
+import {AGENT_TRANSFER_LLM_REQUEST_PROCESSOR} from './processors/agent_transfer_llm_request_processor.js';
+import {BASIC_LLM_REQUEST_PROCESSOR} from './processors/basic_llm_request_processor.js';
+import {CODE_EXECUTION_REQUEST_PROCESSOR} from './processors/code_execution_request_processor.js';
+import {CONTENT_REQUEST_PROCESSOR} from './processors/content_request_processor.js';
+import {IDENTITY_LLM_REQUEST_PROCESSOR} from './processors/identity_llm_request_processor.js';
+import {INSTRUCTIONS_LLM_REQUEST_PROCESSOR} from './processors/instructions_llm_request_processor.js';
+import {REQUEST_CONFIRMATION_LLM_REQUEST_PROCESSOR} from './processors/request_confirmation_llm_request_processor.js';
 import {ReadonlyContext} from './readonly_context.js';
 import {StreamingMode} from './run_config.js';
 
@@ -333,958 +309,6 @@ async function convertToolUnionToTools(
   }
   return await toolUnion.getTools(context);
 }
-
-// --------------------------------------------------------------------------
-// #START Request Processors
-// --------------------------------------------------------------------------
-class BasicLlmRequestProcessor extends BaseLlmRequestProcessor {
-  // eslint-disable-next-line require-yield
-  override async *runAsync(
-    invocationContext: InvocationContext,
-    llmRequest: LlmRequest,
-  ): AsyncGenerator<Event, void, void> {
-    const agent = invocationContext.agent;
-    if (!isLlmAgent(agent)) {
-      return;
-    }
-
-    // set model string, not model instance.
-    llmRequest.model = agent.canonicalModel.model;
-
-    llmRequest.config = {...(agent.generateContentConfig ?? {})};
-    if (agent.outputSchema) {
-      setOutputSchema(llmRequest, agent.outputSchema);
-    }
-
-    if (invocationContext.runConfig) {
-      llmRequest.liveConnectConfig.responseModalities =
-        invocationContext.runConfig.responseModalities;
-      llmRequest.liveConnectConfig.speechConfig =
-        invocationContext.runConfig.speechConfig;
-      llmRequest.liveConnectConfig.outputAudioTranscription =
-        invocationContext.runConfig.outputAudioTranscription;
-      llmRequest.liveConnectConfig.inputAudioTranscription =
-        invocationContext.runConfig.inputAudioTranscription;
-      llmRequest.liveConnectConfig.realtimeInputConfig =
-        invocationContext.runConfig.realtimeInputConfig;
-      llmRequest.liveConnectConfig.enableAffectiveDialog =
-        invocationContext.runConfig.enableAffectiveDialog;
-      llmRequest.liveConnectConfig.proactivity =
-        invocationContext.runConfig.proactivity;
-    }
-  }
-}
-const BASIC_LLM_REQUEST_PROCESSOR = new BasicLlmRequestProcessor();
-
-class IdentityLlmRequestProcessor extends BaseLlmRequestProcessor {
-  // eslint-disable-next-line require-yield
-  override async *runAsync(
-    invocationContext: InvocationContext,
-    llmRequest: LlmRequest,
-  ): AsyncGenerator<Event, void, undefined> {
-    const agent = invocationContext.agent;
-    const si = [`You are an agent. Your internal name is "${agent.name}".`];
-    if (agent.description) {
-      si.push(`The description about you is "${agent.description}"`);
-    }
-    appendInstructions(llmRequest, si);
-  }
-}
-const IDENTITY_LLM_REQUEST_PROCESSOR = new IdentityLlmRequestProcessor();
-
-class InstructionsLlmRequestProcessor extends BaseLlmRequestProcessor {
-  /**
-   * Handles instructions and global instructions for LLM flow.
-   */
-  // eslint-disable-next-line require-yield
-  async *runAsync(
-    invocationContext: InvocationContext,
-    llmRequest: LlmRequest,
-  ): AsyncGenerator<Event, void, void> {
-    const agent = invocationContext.agent;
-    if (!isLlmAgent(agent) || !isLlmAgent(agent.rootAgent)) {
-      return;
-    }
-    const rootAgent: LlmAgent = agent.rootAgent;
-
-    // TODO - b/425992518: unexpected and buggy for performance.
-    // Global instruction should be explicitly scoped.
-    // Step 1: Appends global instructions if set by RootAgent.
-    if (isLlmAgent(rootAgent) && rootAgent.globalInstruction) {
-      const {instruction, requireStateInjection} =
-        await rootAgent.canonicalGlobalInstruction(
-          new ReadonlyContext(invocationContext),
-        );
-      let instructionWithState = instruction;
-      if (requireStateInjection) {
-        instructionWithState = await injectSessionState(
-          instruction,
-          new ReadonlyContext(invocationContext),
-        );
-      }
-      appendInstructions(llmRequest, [instructionWithState]);
-    }
-
-    // Step 2: Appends agent local instructions if set.
-    // TODO - b/425992518: requireStateInjection means user passed a
-    // instruction processor. We need to make it more explicit.
-    if (agent.instruction) {
-      const {instruction, requireStateInjection} =
-        await agent.canonicalInstruction(
-          new ReadonlyContext(invocationContext),
-        );
-      let instructionWithState = instruction;
-      if (requireStateInjection) {
-        instructionWithState = await injectSessionState(
-          instruction,
-          new ReadonlyContext(invocationContext),
-        );
-      }
-      appendInstructions(llmRequest, [instructionWithState]);
-    }
-  }
-}
-const INSTRUCTIONS_LLM_REQUEST_PROCESSOR =
-  new InstructionsLlmRequestProcessor();
-
-class ContentRequestProcessor implements BaseLlmRequestProcessor {
-  // eslint-disable-next-line require-yield
-  async *runAsync(
-    invocationContext: InvocationContext,
-    llmRequest: LlmRequest,
-  ): AsyncGenerator<Event, void, void> {
-    const agent = invocationContext.agent;
-    if (!agent || !isLlmAgent(agent)) {
-      return;
-    }
-
-    if (agent.includeContents === 'default') {
-      // Include full conversation history
-      llmRequest.contents = getContents(
-        invocationContext.session.events,
-        agent.name,
-        invocationContext.branch,
-      );
-    } else {
-      // Include current turn context only (no conversation history).
-      llmRequest.contents = getCurrentTurnContents(
-        invocationContext.session.events,
-        agent.name,
-        invocationContext.branch,
-      );
-    }
-
-    return;
-  }
-}
-const CONTENT_REQUEST_PROCESSOR = new ContentRequestProcessor();
-
-class AgentTransferLlmRequestProcessor extends BaseLlmRequestProcessor {
-  private readonly toolName = 'transfer_to_agent' as const;
-  private readonly tool = new FunctionTool({
-    name: this.toolName,
-    description:
-      'Transfer the question to another agent. This tool hands off control to another agent when it is more suitable to answer the user question according to the agent description.',
-    parameters: z.object({
-      agentName: z.string().describe('the agent name to transfer to.'),
-    }),
-    execute: function (args: {agentName: string}, toolContext?: Context) {
-      if (!toolContext) {
-        throw new Error('toolContext is required.');
-      }
-      toolContext.actions.transferToAgent = args.agentName;
-      return 'Transfer queued';
-    },
-  });
-
-  // eslint-disable-next-line require-yield
-  override async *runAsync(
-    invocationContext: InvocationContext,
-    llmRequest: LlmRequest,
-  ): AsyncGenerator<Event, void, void> {
-    if (!isLlmAgent(invocationContext.agent)) {
-      return;
-    }
-
-    const transferTargets = this.getTransferTargets(invocationContext.agent);
-    if (!transferTargets.length) {
-      return;
-    }
-
-    appendInstructions(llmRequest, [
-      this.buildTargetAgentsInstructions(
-        invocationContext.agent,
-        transferTargets,
-      ),
-    ]);
-
-    const toolContext = new Context({invocationContext});
-    await this.tool.processLlmRequest({toolContext, llmRequest});
-  }
-
-  private buildTargetAgentsInfo(targetAgent: BaseAgent): string {
-    return `
-Agent name: ${targetAgent.name}
-Agent description: ${targetAgent.description}
-`;
-  }
-
-  private buildTargetAgentsInstructions(
-    agent: LlmAgent,
-    targetAgents: BaseAgent[],
-  ): string {
-    let instructions = `
-You have a list of other agents to transfer to:
-
-${targetAgents.map(this.buildTargetAgentsInfo).join('\n')}
-
-If you are the best to answer the question according to your description, you
-can answer it.
-
-If another agent is better for answering the question according to its
-description, call \`${this.toolName}\` function to transfer the
-question to that agent. When transferring, do not generate any text other than
-the function call.
-`;
-
-    if (agent.parentAgent && !agent.disallowTransferToParent) {
-      instructions += `
-Your parent agent is ${agent.parentAgent.name}. If neither the other agents nor
-you are best for answering the question according to the descriptions, transfer
-to your parent agent.
-`;
-    }
-    return instructions;
-  }
-
-  private getTransferTargets(agent: LlmAgent): BaseAgent[] {
-    const targets: BaseAgent[] = [];
-    targets.push(...agent.subAgents);
-
-    if (!agent.parentAgent || !isLlmAgent(agent.parentAgent)) {
-      return targets;
-    }
-
-    if (!agent.disallowTransferToParent) {
-      targets.push(agent.parentAgent);
-    }
-
-    if (!agent.disallowTransferToPeers) {
-      targets.push(
-        ...agent.parentAgent.subAgents.filter(
-          (peerAgent) => peerAgent.name !== agent.name,
-        ),
-      );
-    }
-
-    return targets;
-  }
-}
-const AGENT_TRANSFER_LLM_REQUEST_PROCESSOR =
-  new AgentTransferLlmRequestProcessor();
-
-class RequestConfirmationLlmRequestProcessor extends BaseLlmRequestProcessor {
-  /** Handles tool confirmation information to build the LLM request. */
-  override async *runAsync(
-    invocationContext: InvocationContext,
-  ): AsyncGenerator<Event, void, void> {
-    const agent = invocationContext.agent;
-    if (!isLlmAgent(agent)) {
-      return;
-    }
-    const events = invocationContext.session.events;
-    if (!events || events.length === 0) {
-      return;
-    }
-
-    const requestConfirmationFunctionResponses: {
-      [key: string]: ToolConfirmation;
-    } = {};
-
-    let confirmationEventIndex = -1;
-    // Step 1: Find the FIRST confirmation event authored by user.
-    for (let i = events.length - 1; i >= 0; i--) {
-      const event = events[i];
-      if (event.author !== 'user') {
-        continue;
-      }
-      const responses = getFunctionResponses(event);
-      if (!responses) {
-        continue;
-      }
-
-      let foundConfirmation = false;
-      for (const functionResponse of responses) {
-        if (functionResponse.name !== REQUEST_CONFIRMATION_FUNCTION_CALL_NAME) {
-          continue;
-        }
-        foundConfirmation = true;
-
-        let toolConfirmation = null;
-
-        if (
-          functionResponse.response &&
-          Object.keys(functionResponse.response).length === 1 &&
-          'response' in functionResponse.response
-        ) {
-          toolConfirmation = JSON.parse(
-            functionResponse.response['response'] as string,
-          ) as ToolConfirmation;
-        } else if (functionResponse.response) {
-          toolConfirmation = new ToolConfirmation({
-            hint: functionResponse.response['hint'] as string,
-            payload: functionResponse.response['payload'],
-            confirmed: functionResponse.response['confirmed'] as boolean,
-          });
-        }
-
-        if (functionResponse.id && toolConfirmation) {
-          requestConfirmationFunctionResponses[functionResponse.id] =
-            toolConfirmation;
-        }
-      }
-      if (foundConfirmation) {
-        confirmationEventIndex = i;
-        break;
-      }
-    }
-
-    if (Object.keys(requestConfirmationFunctionResponses).length === 0) {
-      return;
-    }
-
-    // Step 2: Find the system generated FunctionCall event requesting the tool
-    // confirmation
-    for (let i = confirmationEventIndex - 1; i >= 0; i--) {
-      const event = events[i];
-      const functionCalls = getFunctionCalls(event);
-      if (!functionCalls) {
-        continue;
-      }
-
-      const toolsToResumeWithConfirmation: {[key: string]: ToolConfirmation} =
-        {};
-      const toolsToResumeWithArgs: {[key: string]: FunctionCall} = {};
-
-      for (const functionCall of functionCalls) {
-        if (
-          !functionCall.id ||
-          !(functionCall.id in requestConfirmationFunctionResponses)
-        ) {
-          continue;
-        }
-
-        const args = functionCall.args;
-        if (!args || !('originalFunctionCall' in args)) {
-          continue;
-        }
-        const originalFunctionCall = args[
-          'originalFunctionCall'
-        ] as FunctionCall;
-
-        if (originalFunctionCall.id) {
-          toolsToResumeWithConfirmation[originalFunctionCall.id] =
-            requestConfirmationFunctionResponses[functionCall.id];
-          toolsToResumeWithArgs[originalFunctionCall.id] = originalFunctionCall;
-        }
-      }
-      if (Object.keys(toolsToResumeWithConfirmation).length === 0) {
-        continue;
-      }
-
-      // Step 3: Remove the tools that have already been confirmed AND resumed.
-      for (let j = events.length - 1; j > confirmationEventIndex; j--) {
-        const eventToCheck = events[j];
-        const functionResponses = getFunctionResponses(eventToCheck);
-        if (!functionResponses) {
-          continue;
-        }
-
-        for (const fr of functionResponses) {
-          if (fr.id && fr.id in toolsToResumeWithConfirmation) {
-            delete toolsToResumeWithConfirmation[fr.id];
-            delete toolsToResumeWithArgs[fr.id];
-          }
-        }
-        if (Object.keys(toolsToResumeWithConfirmation).length === 0) {
-          break;
-        }
-      }
-
-      if (Object.keys(toolsToResumeWithConfirmation).length === 0) {
-        continue;
-      }
-
-      const toolsList = await agent.canonicalTools(
-        new ReadonlyContext(invocationContext),
-      );
-      const toolsDict = Object.fromEntries(
-        toolsList.map((tool) => [tool.name, tool]),
-      );
-
-      const functionResponseEvent = await handleFunctionCallList({
-        invocationContext: invocationContext,
-        functionCalls: Object.values(toolsToResumeWithArgs),
-        toolsDict: toolsDict,
-        beforeToolCallbacks: agent.canonicalBeforeToolCallbacks,
-        afterToolCallbacks: agent.canonicalAfterToolCallbacks,
-        filters: new Set(Object.keys(toolsToResumeWithConfirmation)),
-        toolConfirmationDict: toolsToResumeWithConfirmation,
-      });
-
-      if (functionResponseEvent) {
-        yield functionResponseEvent;
-      }
-      return;
-    }
-  }
-}
-
-export const REQUEST_CONFIRMATION_LLM_REQUEST_PROCESSOR =
-  new RequestConfirmationLlmRequestProcessor();
-
-/**
- * Processes code execution requests.
- */
-class CodeExecutionRequestProcessor extends BaseLlmRequestProcessor {
-  override async *runAsync(
-    invocationContext: InvocationContext,
-    llmRequest: LlmRequest,
-  ): AsyncGenerator<Event, void, void> {
-    if (!isLlmAgent(invocationContext.agent)) {
-      return;
-    }
-
-    if (!invocationContext.agent.codeExecutor) {
-      return;
-    }
-
-    for await (const event of runPreProcessor(invocationContext, llmRequest)) {
-      yield event;
-    }
-
-    if (!isBaseCodeExecutor(invocationContext.agent.codeExecutor)) {
-      return;
-    }
-
-    for (const content of llmRequest.contents) {
-      const delimeters: [string, string] = invocationContext.agent.codeExecutor
-        .codeBlockDelimiters.length
-        ? invocationContext.agent.codeExecutor.codeBlockDelimiters[0]
-        : ['', ''];
-
-      convertCodeExecutionParts(
-        content,
-        delimeters,
-        invocationContext.agent.codeExecutor.executionResultDelimiters,
-      );
-    }
-  }
-}
-
-/**
- * Map of MIME types to data file utilities
- */
-const DATA_FILE_UTIL_MAP: Record<
-  string,
-  {
-    extension: string;
-    loaderCodeTemplate: string;
-  }
-> = {
-  'text/csv': {
-    extension: '.csv',
-    loaderCodeTemplate: "pd.read_csv('{filename}')",
-  },
-};
-
-/**
- * Helper library for data file exploration
- */
-const DATA_FILE_HELPER_LIB = `
-import pandas as pd
-
-def explore_df(df: pd.DataFrame) -> None:
-  """Prints some information about a pandas DataFrame."""
-
-  with pd.option_context(
-      'display.max_columns', None, 'display.expand_frame_repr', False
-  ):
-    # Print the column names to never encounter KeyError when selecting one.
-    df_dtypes = df.dtypes
-
-    # Obtain information about data types and missing values.
-    df_nulls = (len(df) - df.isnull().sum()).apply(
-        lambda x: f'{x} / {df.shape[0]} non-null'
-    )
-
-    # Explore unique total values in columns using \`.unique()\`.
-    df_unique_count = df.apply(lambda x: len(x.unique()))
-
-    # Explore unique values in columns using \`.unique()\`.
-    df_unique = df.apply(lambda x: crop(str(list(x.unique()))))
-
-    df_info = pd.concat(
-        (
-            df_dtypes.rename('Dtype'),
-            df_nulls.rename('Non-Null Count'),
-            df_unique_count.rename('Unique Values Count'),
-            df_unique.rename('Unique Values'),
-        ),
-        axis=1,
-    )
-    df_info.index.name = 'Columns'
-    print(f"""Total rows: {df.shape[0]}
-Total columns: {df.shape[1]}
-
-{df_info}""")
-`;
-
-/**
- * Processor for code execution responses.
- */
-class CodeExecutionResponseProcessor implements BaseLlmResponseProcessor {
-  /**
-   * Processes the LLM response asynchronously.
-   *
-   * @param invocationContext The invocation context
-   * @param llmResponse The LLM response to process
-   * @returns An async generator yielding events
-   */
-  async *runAsync(
-    invocationContext: InvocationContext,
-    llmResponse: LlmResponse,
-  ): AsyncGenerator<Event, void, unknown> {
-    // Skip if the response is partial (streaming)
-    if (llmResponse.partial) {
-      return;
-    }
-
-    // Run the post-processor with standard generator approach
-    for await (const event of runPostProcessor(
-      invocationContext,
-      llmResponse,
-    )) {
-      yield event;
-    }
-  }
-}
-
-/**
- * The exported response processor instance.
- */
-export const responseProcessor = new CodeExecutionResponseProcessor();
-
-/**
- * Pre-processes the user message by adding the user message to the execution
- * environment.
- *
- * @param invocationContext The invocation context
- * @param llmRequest The LLM request to process
- * @returns An async generator yielding events
- */
-async function* runPreProcessor(
-  invocationContext: InvocationContext,
-  llmRequest: LlmRequest,
-): AsyncGenerator<Event, void, unknown> {
-  const agent = invocationContext.agent;
-
-  if (!isLlmAgent(agent)) {
-    return;
-  }
-
-  const codeExecutor = agent.codeExecutor;
-
-  if (!codeExecutor || !isBaseCodeExecutor(codeExecutor)) {
-    return;
-  }
-
-  if (isBuiltInCodeExecutor(codeExecutor)) {
-    codeExecutor.processLlmRequest(llmRequest);
-    return;
-  }
-
-  if (!codeExecutor.optimizeDataFile) {
-    return;
-  }
-
-  const codeExecutorContext = new CodeExecutorContext(
-    new State(invocationContext.session.state),
-  );
-
-  // Skip if the error count exceeds the max retry attempts
-  if (
-    codeExecutorContext.getErrorCount(invocationContext.invocationId) >=
-    codeExecutor.errorRetryAttempts
-  ) {
-    return;
-  }
-
-  // [Step 1] Extract data files from the session_history and store them in
-  // memory Meanwhile, mutate the inline data file to text part in session
-  // history from all turns
-  const allInputFiles = extractAndReplaceInlineFiles(
-    codeExecutorContext,
-    llmRequest,
-  );
-
-  // [Step 2] Run explore_df code on the data files from the current turn
-  // We only need to explore the new data files because the previous data files
-  // should already be explored and cached in the code execution runtime
-  const processedFileNames = new Set(
-    codeExecutorContext.getProcessedFileNames(),
-  );
-  const filesToProcess = allInputFiles.filter(
-    (f) => !processedFileNames.has(f.name),
-  );
-
-  for (const file of filesToProcess) {
-    const codeStr = getDataFilePreprocessingCode(file);
-
-    // Skip for unsupported file or executor types
-    if (!codeStr) {
-      return;
-    }
-
-    // Emit the code to execute, and add it to the LLM request
-    const codeContent: Content = {
-      role: 'model',
-      parts: [
-        {text: `Processing input file: \`${file.name}\``},
-        buildExecutableCodePart(codeStr),
-      ],
-    };
-
-    llmRequest.contents.push(cloneDeep(codeContent)!);
-
-    yield createEvent({
-      invocationId: invocationContext.invocationId,
-      author: agent.name,
-      branch: invocationContext.branch,
-      content: codeContent,
-    });
-
-    const executionId = getOrSetExecutionId(
-      invocationContext,
-      codeExecutorContext,
-    );
-    const codeExecutionResult = await codeExecutor.executeCode({
-      invocationContext,
-      codeExecutionInput: {
-        code: codeStr,
-        inputFiles: [file],
-        executionId,
-      },
-    });
-
-    // Update the processing results to code executor context
-    codeExecutorContext.updateCodeExecutionResult({
-      invocationId: invocationContext.invocationId,
-      code: codeStr,
-      resultStdout: codeExecutionResult.stdout,
-      resultStderr: codeExecutionResult.stderr,
-    });
-
-    codeExecutorContext.addProcessedFileNames([file.name]);
-
-    // Emit the execution result, and add it to the LLM request
-    const executionResultEvent = await postProcessCodeExecutionResult(
-      invocationContext,
-      codeExecutorContext,
-      codeExecutionResult,
-    );
-
-    yield executionResultEvent;
-    llmRequest.contents.push(cloneDeep(executionResultEvent.content)!);
-  }
-}
-
-/**
- * Post-processes the model response by extracting and executing the first code
- * block.
- *
- * @param invocationContext The invocation context
- * @param llmResponse The LLM response to process
- * @returns An async generator yielding events
- */
-async function* runPostProcessor(
-  invocationContext: InvocationContext,
-  llmResponse: LlmResponse,
-): AsyncGenerator<Event, void, unknown> {
-  const agent = invocationContext.agent;
-
-  if (!isLlmAgent(agent)) {
-    return;
-  }
-
-  const codeExecutor = agent.codeExecutor;
-
-  if (!codeExecutor || !isBaseCodeExecutor(codeExecutor)) {
-    return;
-  }
-
-  if (!llmResponse || !llmResponse.content) {
-    return;
-  }
-
-  if (isBuiltInCodeExecutor(codeExecutor)) {
-    return;
-  }
-
-  const codeExecutorContext = new CodeExecutorContext(
-    new State(invocationContext.session.state),
-  );
-
-  // Skip if the error count exceeds the max retry attempts
-  if (
-    codeExecutorContext.getErrorCount(invocationContext.invocationId) >=
-    codeExecutor.errorRetryAttempts
-  ) {
-    return;
-  }
-
-  // [Step 1] Extract code from the model predict response and truncate the
-  // content to the part with the first code block
-  const responseContent = llmResponse.content;
-  const codeStr = extractCodeAndTruncateContent(
-    responseContent,
-    codeExecutor.codeBlockDelimiters,
-  );
-
-  // Terminal state: no code to execute
-  if (!codeStr) {
-    return;
-  }
-
-  // [Step 2] Executes the code and emit 2 Events for code and execution result
-  yield createEvent({
-    invocationId: invocationContext.invocationId,
-    author: agent.name,
-    branch: invocationContext.branch,
-    content: responseContent,
-  });
-
-  const executionId = getOrSetExecutionId(
-    invocationContext,
-    codeExecutorContext,
-  );
-  const codeExecutionResult = await codeExecutor.executeCode({
-    invocationContext,
-    codeExecutionInput: {
-      code: codeStr,
-      inputFiles: codeExecutorContext.getInputFiles(),
-      executionId,
-    },
-  });
-
-  codeExecutorContext.updateCodeExecutionResult({
-    invocationId: invocationContext.invocationId,
-    code: codeStr,
-    resultStdout: codeExecutionResult.stdout,
-    resultStderr: codeExecutionResult.stderr,
-  });
-
-  yield await postProcessCodeExecutionResult(
-    invocationContext,
-    codeExecutorContext,
-    codeExecutionResult,
-  );
-
-  // [Step 3] Skip processing the original model response
-  // to continue code generation loop
-  llmResponse.content = undefined;
-}
-
-/**
- * Extracts and replaces inline files with file names in the LLM request.
- *
- * @param codeExecutorContext The code executor context
- * @param llmRequest The LLM request to process
- * @returns A list of input files
- */
-function extractAndReplaceInlineFiles(
-  codeExecutorContext: CodeExecutorContext,
-  llmRequest: LlmRequest,
-): File[] {
-  const allInputFiles = codeExecutorContext.getInputFiles();
-  const savedFileNames = new Set(allInputFiles.map((f) => f.name));
-
-  // [Step 1] Process input files from LlmRequest and cache them in CodeExecutor
-  for (let i = 0; i < llmRequest.contents.length; i++) {
-    const content = llmRequest.contents[i];
-
-    // Only process the user message
-    if (content.role !== 'user' || !content.parts) {
-      continue;
-    }
-
-    for (let j = 0; j < content.parts.length; j++) {
-      const part = content.parts[j] as Part;
-      const mimeType = part.inlineData?.mimeType;
-
-      // Skip if the inline data is not supported
-      if (!mimeType || !part.inlineData || !DATA_FILE_UTIL_MAP[mimeType]) {
-        continue;
-      }
-
-      // Replace the inline data file with a file name placeholder
-      const fileName = `data_${i + 1}_${j + 1}${DATA_FILE_UTIL_MAP[mimeType].extension}`;
-
-      part.text = `\nAvailable file: \`${fileName}\`\n`;
-
-      // Add the inline data as input file to the code executor context
-      const file: File = {
-        name: fileName,
-        content: base64Decode(part.inlineData.data!),
-        mimeType,
-      };
-
-      if (!savedFileNames.has(fileName)) {
-        codeExecutorContext.addInputFiles([file]);
-        allInputFiles.push(file);
-      }
-    }
-  }
-
-  return allInputFiles;
-}
-
-/**
- * Gets or sets the execution ID for stateful code execution.
- *
- * @param invocationContext The invocation context
- * @param codeExecutorContext The code executor context
- * @returns The execution ID or undefined if not stateful
- */
-function getOrSetExecutionId(
-  invocationContext: InvocationContext,
-  codeExecutorContext: CodeExecutorContext,
-): string | undefined {
-  const agent = invocationContext.agent;
-
-  if (!isLlmAgent(agent) || !agent.codeExecutor?.stateful) {
-    return undefined;
-  }
-
-  let executionId = codeExecutorContext.getExecutionId();
-
-  if (!executionId) {
-    executionId = invocationContext.session.id;
-    codeExecutorContext.setExecutionId(executionId);
-  }
-
-  return executionId;
-}
-
-/**
- * Post-processes the code execution result and emits an Event.
- *
- * @param invocationContext The invocation context
- * @param codeExecutorContext The code executor context
- * @param codeExecutionResult The code execution result
- * @returns The event with the code execution result
- */
-async function postProcessCodeExecutionResult(
-  invocationContext: InvocationContext,
-  codeExecutorContext: CodeExecutorContext,
-  codeExecutionResult: CodeExecutionResult,
-): Promise<Event> {
-  if (!invocationContext.artifactService) {
-    throw new Error('Artifact service is not initialized.');
-  }
-
-  const resultContent: Content = {
-    role: 'model',
-    parts: [buildCodeExecutionResultPart(codeExecutionResult)],
-  };
-
-  const eventActions = createEventActions({
-    stateDelta: codeExecutorContext.getStateDelta(),
-  });
-
-  // Handle code execution error retry
-  if (codeExecutionResult.stderr) {
-    codeExecutorContext.incrementErrorCount(invocationContext.invocationId);
-  } else {
-    codeExecutorContext.resetErrorCount(invocationContext.invocationId);
-  }
-
-  // Handle output files
-  for (const outputFile of codeExecutionResult.outputFiles) {
-    const version = await invocationContext.artifactService.saveArtifact({
-      appName: invocationContext.appName || '',
-      userId: invocationContext.userId || '',
-      sessionId: invocationContext.session.id,
-      filename: outputFile.name,
-      artifact: {
-        inlineData: {data: outputFile.content, mimeType: outputFile.mimeType},
-      },
-    });
-
-    eventActions.artifactDelta[outputFile.name] = version;
-  }
-
-  return createEvent({
-    invocationId: invocationContext.invocationId,
-    author: invocationContext.agent.name,
-    branch: invocationContext.branch,
-    content: resultContent,
-    actions: eventActions,
-  });
-}
-
-/**
- * Returns the code to explore the data file.
- *
- * @param file The file to explore
- * @returns The code to explore the data file or undefined if not supported
- */
-function getDataFilePreprocessingCode(file: File): string | undefined {
-  /**
-   * Gets a normalized file name.
-   *
-   * @param fileName The file name to normalize
-   * @returns The normalized file name
-   */
-  function getNormalizedFileName(fileName: string): string {
-    const [varName] = fileName.split('.');
-
-    // Replace non-alphanumeric characters with underscores
-    let normalizedName = varName.replace(/[^a-zA-Z0-9_]/g, '_');
-
-    // If the filename starts with a digit, prepend an underscore
-    if (/^\d/.test(normalizedName)) {
-      normalizedName = '_' + normalizedName;
-    }
-
-    return normalizedName;
-  }
-
-  if (!DATA_FILE_UTIL_MAP[file.mimeType]) {
-    return undefined;
-  }
-
-  const varName = getNormalizedFileName(file.name);
-  const loaderCode = DATA_FILE_UTIL_MAP[
-    file.mimeType
-  ].loaderCodeTemplate.replace('{filename}', file.name);
-
-  return `
-${DATA_FILE_HELPER_LIB}
-
-# Load the dataframe.
-${varName} = ${loaderCode}
-
-# Use \`explore_df\` to guide my analysis.
-explore_df(${varName})
-`;
-}
-
-const CODE_EXECUTION_REQUEST_PROCESSOR = new CodeExecutionRequestProcessor();
-
-// --------------------------------------------------------------------------
-// #END RequesBaseCodeExecutort Processors
-// --------------------------------------------------------------------------
 
 /**
  * A unique symbol to identify ADK agent classes.
