+++ core/src/agents/processors/request_confirmation_llm_request_processor.ts
@@ -1,0 +1,181 @@
+/**
+ * @license
+ * Copyright 2025 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import {FunctionCall} from '@google/genai';
+import {
+  Event,
+  getFunctionCalls,
+  getFunctionResponses,
+} from '../../events/event.js';
+import {ToolConfirmation} from '../../tools/tool_confirmation.js';
+import {
+  REQUEST_CONFIRMATION_FUNCTION_CALL_NAME,
+  handleFunctionCallList,
+} from '../functions.js';
+import {InvocationContext} from '../invocation_context.js';
+import {isLlmAgent} from '../llm_agent.js';
+import {ReadonlyContext} from '../readonly_context.js';
+import {BaseLlmRequestProcessor} from './base_llm_processor.js';
+
+export class RequestConfirmationLlmRequestProcessor extends BaseLlmRequestProcessor {
+  /** Handles tool confirmation information to build the LLM request. */
+  override async *runAsync(
+    invocationContext: InvocationContext,
+  ): AsyncGenerator<Event, void, void> {
+    const agent = invocationContext.agent;
+    if (!isLlmAgent(agent)) {
+      return;
+    }
+    const events = invocationContext.session.events;
+    if (!events || events.length === 0) {
+      return;
+    }
+
+    const requestConfirmationFunctionResponses: {
+      [key: string]: ToolConfirmation;
+    } = {};
+
+    let confirmationEventIndex = -1;
+    // Step 1: Find the FIRST confirmation event authored by user.
+    for (let i = events.length - 1; i >= 0; i--) {
+      const event = events[i];
+      if (event.author !== 'user') {
+        continue;
+      }
+      const responses = getFunctionResponses(event);
+      if (!responses) {
+        continue;
+      }
+
+      let foundConfirmation = false;
+      for (const functionResponse of responses) {
+        if (functionResponse.name !== REQUEST_CONFIRMATION_FUNCTION_CALL_NAME) {
+          continue;
+        }
+        foundConfirmation = true;
+
+        let toolConfirmation = null;
+
+        if (
+          functionResponse.response &&
+          Object.keys(functionResponse.response).length === 1 &&
+          'response' in functionResponse.response
+        ) {
+          toolConfirmation = JSON.parse(
+            functionResponse.response['response'] as string,
+          ) as ToolConfirmation;
+        } else if (functionResponse.response) {
+          toolConfirmation = new ToolConfirmation({
+            hint: functionResponse.response['hint'] as string,
+            payload: functionResponse.response['payload'],
+            confirmed: functionResponse.response['confirmed'] as boolean,
+          });
+        }
+
+        if (functionResponse.id && toolConfirmation) {
+          requestConfirmationFunctionResponses[functionResponse.id] =
+            toolConfirmation;
+        }
+      }
+      if (foundConfirmation) {
+        confirmationEventIndex = i;
+        break;
+      }
+    }
+
+    if (Object.keys(requestConfirmationFunctionResponses).length === 0) {
+      return;
+    }
+
+    // Step 2: Find the system generated FunctionCall event requesting the tool
+    // confirmation
+    for (let i = confirmationEventIndex - 1; i >= 0; i--) {
+      const event = events[i];
+      const functionCalls = getFunctionCalls(event);
+      if (!functionCalls) {
+        continue;
+      }
+
+      const toolsToResumeWithConfirmation: {[key: string]: ToolConfirmation} =
+        {};
+      const toolsToResumeWithArgs: {[key: string]: FunctionCall} = {};
+
+      for (const functionCall of functionCalls) {
+        if (
+          !functionCall.id ||
+          !(functionCall.id in requestConfirmationFunctionResponses)
+        ) {
+          continue;
+        }
+
+        const args = functionCall.args;
+        if (!args || !('originalFunctionCall' in args)) {
+          continue;
+        }
+        const originalFunctionCall = args[
+          'originalFunctionCall'
+        ] as FunctionCall;
+
+        if (originalFunctionCall.id) {
+          toolsToResumeWithConfirmation[originalFunctionCall.id] =
+            requestConfirmationFunctionResponses[functionCall.id];
+          toolsToResumeWithArgs[originalFunctionCall.id] = originalFunctionCall;
+        }
+      }
+      if (Object.keys(toolsToResumeWithConfirmation).length === 0) {
+        continue;
+      }
+
+      // Step 3: Remove the tools that have already been confirmed AND resumed.
+      for (let j = events.length - 1; j > confirmationEventIndex; j--) {
+        const eventToCheck = events[j];
+        const functionResponses = getFunctionResponses(eventToCheck);
+        if (!functionResponses) {
+          continue;
+        }
+
+        for (const fr of functionResponses) {
+          if (fr.id && fr.id in toolsToResumeWithConfirmation) {
+            delete toolsToResumeWithConfirmation[fr.id];
+            delete toolsToResumeWithArgs[fr.id];
+          }
+        }
+        if (Object.keys(toolsToResumeWithConfirmation).length === 0) {
+          break;
+        }
+      }
+
+      if (Object.keys(toolsToResumeWithConfirmation).length === 0) {
+        continue;
+      }
+
+      const toolsList = await agent.canonicalTools(
+        new ReadonlyContext(invocationContext),
+      );
+      const toolsDict = Object.fromEntries(
+        toolsList.map((tool) => [tool.name, tool]),
+      );
+
+      const functionResponseEvent = await handleFunctionCallList({
+        invocationContext: invocationContext,
+        functionCalls: Object.values(toolsToResumeWithArgs),
+        toolsDict: toolsDict,
+        beforeToolCallbacks: agent.canonicalBeforeToolCallbacks,
+        afterToolCallbacks: agent.canonicalAfterToolCallbacks,
+        filters: new Set(Object.keys(toolsToResumeWithConfirmation)),
+        toolConfirmationDict: toolsToResumeWithConfirmation,
+      });
+
+      if (functionResponseEvent) {
+        yield functionResponseEvent;
+      }
+      return;
+    }
+  }
+}
+
+export const REQUEST_CONFIRMATION_LLM_REQUEST_PROCESSOR =
+  new RequestConfirmationLlmRequestProcessor();
